// _backup_/lib/input/types.ts
export type TgdTypeKeyAlphabetLow =
    | "a"
    | "b"
    | "c"
    | "d"
    | "e"
    | "f"
    | "g"
    | "h"
    | "i"
    | "j"
    | "k"
    | "l"
    | "m"
    | "n"
    | "o"
    | "p"
    | "q"
    | "r"
    | "s"
    | "t"
    | "u"
    | "v"
    | "w"
    | "x"
    | "y"
    | "z"

export type TgdTypeKeyAlphabetUp =
    | "A"
    | "B"
    | "C"
    | "D"
    | "E"
    | "F"
    | "G"
    | "H"
    | "I"
    | "J"
    | "K"
    | "L"
    | "M"
    | "N"
    | "O"
    | "P"
    | "Q"
    | "R"
    | "S"
    | "T"
    | "U"
    | "V"
    | "W"
    | "X"
    | "Y"
    | "Z"

export type TgdTypeKeyAlphabet = TgdTypeKeyAlphabetLow | TgdTypeKeyAlphabetUp

export type TgdTypeKeySpecial = "Unidentified"

export type TgdTypeKeyModifier =
    | "Alt"
    | "AltGraph"
    | "CapsLock"
    | "Control"
    | "Fn"
    | "FnLock"
    | "Hyper"
    | "Meta"
    | "NumLock"
    | "ScrollLock"
    | "Shift"
    | "Super"
    | "Symbol"
    | "SymbolLock"

export type TgdTypeKeyWhiteSpace = "Enter" | "Tab" | " "

export type TgdTypeKeyNavigation =
    | "ArrowDown"
    | "ArrowLeft"
    | "ArrowRight"
    | "ArrowUp"
    | "End"
    | "Home"
    | "PageDown"
    | "PageUp"

export type TgdTypeKeyEditing =
    | "Backspace"
    | "Clear"
    | "Copy"
    | "CrSel"
    | "Cut"
    | "Delete"
    | "EraseOf"
    | "ExSel"
    | "Insert"
    | "Paste"
    | "Redo"
    | "Undo"

export type TgdTypeKeyUI =
    | "Accept"
    | "Again"
    | "Attn"
    | "Cancel"
    | "ContextMenu"
    | "Escape"
    | "Execute"
    | "Find"
    | "Finish"
    | "Help"
    | "Pause"
    | "Play"
    | "Props"
    | "Select"
    | "ZoomIn"
    | "ZoomOut"

export type TgdTypeKeyDevice =
    | "BrightnessDown"
    | "BrightnessUp"
    | "Eject"
    | "LogOff"
    | "Power"
    | "PowerOff"
    | "PrintScreen"
    | "Hibernate"
    | "Standby"
    | "WakeUp"

export type TgdTypeKeyFunction =
    | "F1"
    | "F2"
    | "F3"
    | "F4"
    | "F5"
    | "F6"
    | "F7"
    | "F8"
    | "F9"
    | "F10"
    | "F11"
    | "F12"
    | "F13"
    | "F14"
    | "F15"
    | "F16"
    | "F17"
    | "F18"
    | "F19"
    | "F20"
    | "Soft1"
    | "Soft2"
    | "Soft3"
    | "Soft4"

export type TgdTypeKeyPhone =
    | "AppSwitch"
    | "Call"
    | "Camera"
    | "CameraFocus"
    | "EndCall"
    | "GoBack"
    | "GoHome"
    | "HeadsetHook"
    | "LastNumberRedial"
    | "Notification"
    | "MannerMode"
    | "VoiceDial"

export type TgdTypeKeyMultimedia =
    | "ChannelDown"
    | "ChannelUp"
    | "MediaFastForward"
    | "MediaPause"
    | "MediaPlay"
    | "MediaPlayPause"
    | "MediaRecord"
    | "MediaRewind"
    | "MediaStop"
    | "MediaTrackNext"
    | "MediaTrackPrevious"

export type TgdTypeKeyAudio =
    | "AudioBalanceLeft"
    | "AudioBalanceRight"
    | "AudioBassDown"
    | "AudioBassBoostDown"
    | "AudioBassBoostToggle"
    | "AudioBassBoostUp"
    | "AudioBassUp"
    | "AudioFaderFront"
    | "AudioFaderRear"
    | "AudioSurroundModeNext"
    | "AudioTrebleDown"
    | "AudioTrebleUp"
    | "AudioVolumeDown"
    | "AudioVolumeMute"
    | "AudioVolumeUp"
    | "MicrophoneToggle"
    | "MicrophoneVolumeDown"
    | "MicrophoneVolumeMute"
    | "MicrophoneVolumeUp"

export type TgdTypeKeyTV =
    | "TV"
    | "TV3DMode"
    | "TVAntennaCable"
    | "TVAudioDescription"
    | "TVAudioDescriptionMixDown"
    | "TVAudioDescriptionMixUp"
    | "TVContentsMenu"
    | "TVDataService"
    | "TVInput"
    | "TVInputComponent1"
    | "TVInputComponent2"
    | "TVInputComposite1"
    | "TVInputComposite2"
    | "TVInputHDMI1"
    | "TVInputHDMI2"
    | "TVInputHDMI3"
    | "TVInputHDMI4"
    | "TVInputVGA1"
    | "TVMediaContext"
    | "TVNetwork"
    | "TVNumberEntry"
    | "TVPower"
    | "TVRadioService"
    | "TVSatellite"
    | "TVSatelliteBS"
    | "TVSatelliteCS"
    | "TVSatelliteToggle"
    | "TVTerrestrialAnalog"
    | "TVTerrestrialDigital"
    | "TVTimer"

export type TgdTypeKeyDocument =
    | "Close"
    | "New"
    | "Open"
    | "Print"
    | "Save"
    | "SpellCheck"
    | "MailForward"
    | "MailReply"
    | "MailSend"

export type TgdTypeKeyApplication =
    | "LaunchCalculator"
    | "LaunchCalendar"
    | "LaunchContacts"
    | "LaunchMail"
    | "LaunchMediaPlayer"
    | "LaunchMusicPlayer"
    | "LaunchMyComputer"
    | "LaunchPhone"
    | "LaunchScreenSaver"
    | "LaunchSpreadsheet"
    | "LaunchWebBrowser"
    | "LaunchWebCam"
    | "LaunchWordProcessor"
    | "LaunchApplication1"
    | "LaunchApplication2"
    | "LaunchApplication3"
    | "LaunchApplication4"
    | "LaunchApplication5"
    | "LaunchApplication6"
    | "LaunchApplication7"
    | "LaunchApplication8"
    | "LaunchApplication9"
    | "LaunchApplication10"
    | "LaunchApplication11"
    | "LaunchApplication12"
    | "LaunchApplication13"
    | "LaunchApplication14"
    | "LaunchApplication15"
    | "LaunchApplication16"

export type TgdTypeKeyBrowser =
    | "BrowserBack"
    | "BrowserFavorites"
    | "BrowserForward"
    | "BrowserHome"
    | "BrowserRefresh"
    | "BrowserSearch"
    | "BrowserStop"

export type TgdTypeKeyNumericKeypad =
    | "Decimal"
    | "Key11"
    | "Key12"
    | "Multiply"
    | "Add"
    | "Clear"
    | "Divide"
    | "Subtract"
    | "Separator"
    | "0"
    | "1"
    | "2"
    | "3"
    | "4"
    | "5"
    | "6"
    | "7"
    | "8"
    | "9"

export type TgdTypeKeySymbols =
    | "/"
    | "*"
    | "-"
    | "+"
    | "."
    | "!"
    | "@"
    | "#"
    | "$"
    | "%"
    | "^"
    | "&"
    | "("
    | ")"
    | "_"
    | "="
    | "{"
    | "["
    | "}"
    | "]"
    | ":"
    | ";"
    | "'"
    | '"'
    | "<"
    | ","
    | ">"
    | "\\"
    | "|"
    | "~"
    | " "

export type TgdTypeKey =
    | TgdTypeKeySymbols
    | TgdTypeKeyEditing
    | TgdTypeKeyFunction
    | TgdTypeKeyNavigation
    | TgdTypeKeyNumericKeypad
    | TgdTypeKeyWhiteSpace
    | TgdTypeKeyModifier
    | TgdTypeKeyUI
    | TgdTypeKeySpecial
    | TgdTypeKeyPhone
    | TgdTypeKeyAlphabet
    | TgdTypeKeyDevice
    | TgdTypeKeyMultimedia
    | TgdTypeKeyAudio
    | TgdTypeKeyTV
    | TgdTypeKeyDocument
    | TgdTypeKeyApplication
    | TgdTypeKeyBrowser